move log generating to own class
